feat: 지하철 요금 정책 추가

+++ src/main/java/nextstep/subway/application/dto/response/LineResponse.java
@@ -32,7 +32,7 @@ public class LineResponse {
     private String name;
 
     private String color;
-    private Long additionalFee;
+    private Integer additionalFee;
 
     private List<StationDto> stations;
